♻️ refactor: video관련 로직 분리
